agents/agent-engine/coding/webhook/requirements.txt
flask==3.1.0
vertexai==1.71.1
google-cloud-firestore==2.20.1
python-dotenv==1.0.1
